# src/TypeScriptSchemaCompare.App/src/app/@api/models/get-type-script-model-by-id-response.ts
/* tslint:disable */
import { TypeScriptModelDto } from './type-script-model-dto';
export interface GetTypeScriptModelByIdResponse {
  typeScriptModel?: TypeScriptModelDto;
  validationErrors?: Array<string>;
}