# src/TypeScriptSchemaCompare.App/src/app/@api/models/get-type-script-models-page-response.ts
/* tslint:disable */
import { TypeScriptModelDto } from './type-script-model-dto';
export interface GetTypeScriptModelsPageResponse {
  entities?: Array<TypeScriptModelDto>;
  length?: number;
  validationErrors?: Array<string>;
}