# src/TypeScriptSchemaCompare.App/src/app/@api/models/remove-type-script-model-response.ts
/* tslint:disable */
import { TypeScriptModelDto } from './type-script-model-dto';
export interface RemoveTypeScriptModelResponse {
  typeScriptModel?: TypeScriptModelDto;
  validationErrors?: Array<string>;
}